Move to a better dir and fix to new FSDP name.
